+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:agnes_app/routes/home_page.dart';
+import 'package:agnes_app/routes/login_page.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -23,120 +24,6 @@ class MyApp extends StatelessWidget {
         // This is the theme of your application.
         primarySwatch: Colors.grey,
         scaffoldBackgroundColor: Colors.white,
-      ),
-    );
-  }
-}
-
-class LoginPage extends StatefulWidget {
-  const LoginPage({Key? key, required this.title}) : super(key: key);
-
-  // This widget is the home page of your application. It is stateful, meaning
-  // that it has a State object (defined below) that contains fields that affect
-  // how it looks.
-
-  final String title;
-
-  @override
-  State<LoginPage> createState() => _LoginPageState();
-}
-
-class _LoginPageState extends State<LoginPage> {
-  @override
-  Widget build(BuildContext context) {
-    double width = MediaQuery.of(context).size.width;
-    double height = MediaQuery.of(context).size.height;
-
-    TextEditingController nameController = TextEditingController();
-    TextEditingController passwordController = TextEditingController();
-
-    return Scaffold(
-      body: SizedBox(
-        height: height,
-        width: width,
-        child: Padding(
-          padding: const EdgeInsets.all(5),
-          child: ListView(
-            children: <Widget>[
-              Container(
-                alignment: Alignment.center,
-                padding: const EdgeInsets.all(20),
-                child: const Text(
-                  'Agnes',
-                  style: TextStyle(
-                    color: Color(0xff78af9f),
-                    fontWeight: FontWeight.w500,
-                    fontSize: 40,
-                  ),
-                ),
-              ),
-              Container(
-                padding: const EdgeInsets.all(10),
-                child: TextField(
-                  controller: nameController,
-                  decoration: const InputDecoration(
-                    border: OutlineInputBorder(),
-                    labelText: 'E-mail',
-                  ),
-                ),
-              ),
-              Container(
-                padding: const EdgeInsets.fromLTRB(10, 10, 10, 0),
-                child: TextField(
-                  obscureText: true,
-                  controller: passwordController,
-                  decoration: const InputDecoration(
-                    border: OutlineInputBorder(),
-                    labelText: 'Senha',
-                  ),
-                ),
-              ),
-              TextButton(
-                onPressed: () {
-                  //forgot password screen
-                },
-                child: const Text(
-                  'Esqueceu a senha?',
-                ),
-              ),
-              Container(
-                height: 50,
-                padding: const EdgeInsets.fromLTRB(10, 0, 10, 0),
-                child: ElevatedButton(
-                  child: const Text(
-                    'Entrar',
-                    style: TextStyle(
-                      fontSize: 15,
-                      color: Color(0xff363639),
-                    ),
-                  ),
-                  onPressed: () {
-                    Navigator.of(context)
-                        .pushNamedAndRemoveUntil('/home', (route) => false);
-                  },
-                ),
-              ),
-              Row(
-                children: <Widget>[
-                  const Text('Não tem uma conta?'),
-                  TextButton(
-                    child: const Text(
-                      'Criar',
-                      style: TextStyle(
-                        fontSize: 20,
-                        color: Color(0xff78af9f),
-                      ),
-                    ),
-                    onPressed: () {
-                      //signup screen
-                    },
-                  )
-                ],
-                mainAxisAlignment: MainAxisAlignment.center,
-              ),
-            ],
-          ),
-        ),
       ),
     );
   }
